# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_kiarashAlizadeh_herotech_internal_dto.CreateGuildRequest:
    properties:
      daily_limit:
        type: integer
      name:
        type: string
    type: object
...
  github_com_kiarashAlizadeh_herotech_internal_dto.DepositGoldRequest:
    properties:
      amount:
        type: integer
    type: object
  github_com_kiarashAlizadeh_herotech_internal_dto.GuildInventoryResponse:
    properties:
      listed_items:
        description: Items currently up for sale (available or in auction)
        items:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.ItemResponse'
        type: array
      purchased_items:
        description: Items successfully bought and owned by the guild
        items:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.ItemResponse'
        type: array
    type: object
  github_com_kiarashAlizadeh_herotech_internal_dto.GuildResponse:
    properties:
      created_at:
        type: string
      daily_limit:
        type: integer
      gold_balance:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
  github_com_kiarashAlizadeh_herotech_internal_dto.ItemResponse:
    properties:
      base_price:
        type: integer
      created_at:
        type: string
      id:
        type: string
      list_price:
        type: integer
      name:
        type: string
      owner_id:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  ? github_com_kiarashAlizadeh_herotech_internal_dto.PaginatedResponse-github_com_kiarashAlizadeh_herotech_internal_dto_GuildResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.GuildResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.PaginationMeta'
    type: object
...
  github_com_kiarashAlizadeh_herotech_internal_dto.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_kiarashAlizadeh_herotech_internal_dto.WalletSummaryResponse:
    properties:
      available_balance:
        type: integer
      reserved_amount:
        type: integer
      total_balance:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@Aethoria.com
    name: API Support
  description: Core infrastructure backend for managing Guild Wallets, Item Trading,
    and Legendary Auctions.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DRAGON MARKET PLACE API
  version: "1.0"
paths:
...
  /guilds:
    get:
      description: Retrieves all guild profiles present in the ledger system
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.PaginatedResponse-github_com_kiarashAlizadeh_herotech_internal_dto_GuildResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all registered guilds
      tags:
      - guilds
    post:
      consumes:
      - application/json
      description: Creates a new guild profile with a defined daily spending limit
      parameters:
      - description: Guild payload specs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.CreateGuildRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.GuildResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new guild
      tags:
      - guilds
  /guilds/{id}/inventory:
    get:
      description: Retrieves segregated asset lists highlighting both listed items
        and purchased items
      parameters:
      - description: Guild UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.GuildInventoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch guild inventory profile
      tags:
      - guilds
  /guilds/{id}/wallet:
    get:
      description: Retrieves financial metrics including total balance, reserved funds,
        and liquid available gold
      parameters:
      - description: Guild UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.WalletSummaryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch guild wallet summary
      tags:
      - guilds
  /guilds/{id}/wallet/deposit:
    post:
      consumes:
      - application/json
      description: Increments a specific guild's gold balance ledger sequence
      parameters:
      - description: Guild UUID
        in: path
        name: id
        required: true
        type: string
      - description: Deposit amount parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.DepositGoldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kiarashAlizadeh_herotech_internal_dto.GuildResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deposit gold into guild wallet
      tags:
      - guilds
...
swagger: "2.0"